22/01/24
Created a separate 10x10 array playerBoard as integer. This will be easier to store information such as hits, misses, ships and empty. I have done the same for the opponent.

    Continuation of SetShips array. setShips has a two nested for loops to clear the playerBoard, and two to set the playerShips to be coloured gray on the player's grid.
    Added getXCoords. This is a function designed to take a sender name (eg. picOppA2) and return an x value.
    Added getYCoords. This is a function which serves a very similar purpose to getXCoords

Error time! I have one regarding pictureboxes, when i am trying to access them based on sender data in the onclick method.  To resolve this i will need to make sure i have correctly accessed the Sender details and sent the correct data to the functions.

For next session -> Error resolution and moving forward with how the game works