28/01/24
The goal for today is to fix the event when the opponent's grid is clicked, that only Y values work. And also checking for whether the spot has been taken by a ship or not.

Using a breakpoint, i was able to identify that the getXCoord (from the letters) was checking for a different character in the string. I have now changed it to the correct value. This has resolved the issue

Through simple testing, I was able to identify that the game is looking at the coordinates the wrong way for the ships. Instead of C1, it was A3. I have changed around the x and y in the check if taken if statement. This has resolved the issue. It is slightly problematic as you reference grid squares by y value first.

added two new variables, playerShips and opponentShips. These will store the number of unhit ship sections for each player. Once either of these hits 0, the game is over.

Added a basic function isGameOver (below). This will check if the game has been won, and return one of three values. 0 for game not over, 1 for computer win and 2 for player win.

Added a subroutine for the computer in easy mode. The easy mode computer generates two random numbers between 1 and 10 and uses each to guess a random coordinate. - I had initially forgotten to add an exit to the while loop so the program stopped responding. Exits have now been added.

Added a function for generating a random integer from 1 to upper. This is done by passing in the upper value of the range.

On running the program at this time, the already guessed square method is not working as it is supposed to. The opponent's ships are not showing up as hit. The player grid's hits are not showing up correctly

I have noticed in the setShips method it clears the playerGrid, but not the opponentGrid. It now clears both so i now hope it works.

Less coordinates now say they've been guessed incorrectly.

Next steps. Properly fix duplicate guess method.