3/03/24
As of today

Easy computer mode works perfectly. Hard and Impossible do not.

Today
in Impossible and Hard, replaced non-functioning code with easy. I will write these again differently.
Created an array to hold all of the coordinates for each ship. This way, impossible can just have a counter and do the next move
Worked on impossible move. Started with a move counter, and instructions in what to do each turn via case statement. Then added code to assign each ship to an array in the placeShips subroutine which broke it. Works fine when removed. tried putting. the code in the setships sub instead but whereever i [put it i get a Object reference not set to an instance of an object and it is driving me mad because i dont understand why.,
Next steps
New form with high scores and play again
Stuff each ship sunk notifications,  it's not working.
fix impossible mode
Code Comments
Hard? Not sure how to implement this one