11/03/24
Plan for today: frmGameEnd, (statistics, save score to file, etc) and add some more things to the frmGame.

For frmGameEnd, i need to implement a scoring system.
For this i think a system of starting at 150 points, -1 per miss, +20 per hit, -50 for a loss and +100 for a win. This way the maximum score is 540 and minimum is 0. From there we can multiply by -1 and + 100 to get a value for a progress bar. The progress bar shows the player's progress in sinking the opponent's ships