07/04/24
Centered text in turn thing
I had a thought on how to do the shipsunk notifications and it utilises setting the ships more delibarately on the board (individual integers on the board). so will try that out
45678,
This method works, now i need to render the ships on the board correctly as the previous 1 value is now nonexistant
I now need to fix the progress bar to be more reliant on the ships remaining stat but either way it now doesn't work
I think i should also make ship a record, so it also has a sunk property so it doesn't repeatedly show the you have sunk this ship notificaiton.
Also, i don't have to call the checkIfSUnk every move, just on hits after the second one. calling it from the second is a bit complicated, so instead i will just do after every hit. this saves on performance and means that it doesn't check every single time.
The new implementation of the ships on the board should also mean it is easier to make the impossible game mode again.
Last thing i did today is added a progress bar to each side. This will update each time a ship is sunk.